refactor(#41): 채팅방 조회 서비스 코드 품질 개선

+++ src/main/java/com/example/RealMatch/chat/presentation/dto/response/ChatRoomCardResponse.java
@@ -15,7 +15,7 @@ public record ChatRoomCardResponse(
         String lastMessagePreview,
         ChatMessageType lastMessageType,
         LocalDateTime lastMessageAt,
-        int unreadCount,
+        long unreadCount,
         ChatRoomTab tabCategory
 ) {
 }
